requirements.txt
caldav==1.3.6
python-dotenv==1.0.0
requests==2.31.0
icalendar==5.0.11
pytz==2023.3
python-dateutil==2.8.2
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6 